+++ lib/Path.rb
@@ -1,74 +1,14 @@
+require 'rubygems'
+require 'extensions/kernel'
 require 'pathname'
 
 module WebBlocks
   
   module Path
     
-    module Source
-      
-      def package_dir package
-        ::WebBlocks::Path.from_root_to config[:package][:dir], config[:package][package][:dir]
-      end
-      
-    end
-    
-    module Temporary_Build
-      
-      def tmp_build_dir
-        ::WebBlocks::Path.from_root_to config[:build][:dir_tmp]
-      end
-      
-      def tmp_css_build_dir
-        from_tmp_build_dir_to config[:build][:css][:dir]
-      end
-      
-      def tmp_css_build_file
-        from_tmp_build_dir_to config[:build][:css][:dir], config[:build][:css][:name]
-      end
-      
-      def tmp_css_build_file_ie
-        from_tmp_build_dir_to config[:build][:css][:dir], config[:build][:css][:name_ie]
-      end
-      
-      def tmp_img_build_dir
-        from_tmp_build_dir_to config[:build][:img][:dir]
-      end
-      
-      def tmp_js_build_dir
-        from_tmp_build_dir_to config[:build][:js][:dir]
-      end
-      
-      def tmp_js_build_file
-        from_tmp_build_dir_to config[:build][:js][:dir], config[:build][:js][:name]
-      end
-      
-      def tmp_js_build_file_ie
-        from_tmp_build_dir_to config[:build][:js][:dir], config[:build][:js][:name_ie]
-      end
-      
-      def tmp_js_build_script_dir
-        from_tmp_build_dir_to config[:build][:js][:dir], config[:build][:js][:name_script_dir]
-      end
-      
-      def from_tmp_build_dir_to *args
-        args.unshift tmp_build_dir
-        ::WebBlocks::Path.from_arr_to(args)
-      end
-      
-    end
-    
-    module Build
-      
-      def build_dir
-        ::WebBlocks::Path.from_root_to config[:build][:dir]
-      end
-      
-      def from_build_dir_to *args
-        args.unshift build_dir
-        ::WebBlocks::Path.from_arr_to(args)
-      end
-      
-    end
+    require_relative 'Path/Build'
+    require_relative 'Path/Source'
+    require_relative 'Path/Temporary_Build'
     
     def self.to *args
       ::WebBlocks::Path.from_arr_to(args)
